renamed json_rpc -> rpc
